Fix flex-container layout
The introduction of a sidebar compromised the header's dimension. The
sidebar took up more than the entire viewport, causing the other two
flex items to take as little space as possible, since it had
`flex-grow: 1`. This change uses `flex-shrink: 0` to ensure that flex
items do not shrink smaller than the size of their children.