Patch over issue with pip-compile and git URLs
This is a temporary hack until our django-logging-json fixed gets
patched upstream. There are some tickets relating to this in pip-compile:
https://github.com/jazzband/pip-tools/issues/272

Honestly, I dont have the time bandwidth to investigate for this issue. The
work-around is to not really take advantage of pip-compile for the time being
which I think is a fine trade-off to get stuff working.