expand CI script output, add sass-linter
The sass-linter logic in particular requires some explanation since it looks
really dumb. Basically, the first call (with the tail -f /dev/null) at the end
is a hack to keep the docker image running in the background. From there we copy
the current git repo code to the remote docker container. Finally, we execute
the sass-lint script -- if that call fails it should return an non-zero exit
code and crash the script