Remove sass-lint step from CI
Currently invocation in this way is not supported. Need to mess with the
CI docker image to include these tools. At first I was hesitant to do that, but
I will likely do that in a future PR. For now though, Ive folded the linting
logic into the `dev-go` kickoff script.