Add statistics_tags to built-in tag list
This is primarily a user-experience improvement.  In order for users
to be able to include our statistics tags in page content, we need to
"load" them first.  This means their page must either include the `{%
load statistics_tags %}` directive, or we have to add it to these
builtins.  I'd rather not burden users with remembering to do the
former for every post.

There may eventually be a better way to handle this?  This can make it
somewhat surprising for developers that these statistics tags are
always available in all templates, but right now that may be a
necessary consequence of saving the post-writers some work.