Allow devs to register statistics functions
Two functions here: one that you can use as a decorator on any
function to place that function into a statistics registry to be
usable elsewhere on the site.  These functions are basically template
tags as well but they exist in this special category.

Also, a `get_stats` function for retrieving the set of these functions
and their names.