Peg docker digest hash for WWW image
Utilize the digest hash to pull the www container image from the quay.io repo.
Without this we would be pulling from either the `latest` tag by default.
Im not 100% convinced from the documentation if this is more 'secure'.

Seems like the digest is the hash of the image manifest and not a hash of
all the layers. Could be wrong there ... but anywhoo the point is we should
peg to a hash regardless. They are more reliable for versioning versus,
trying to manage tags ourselves.