Add initial implementation of StatisticsBlock
This block enables, I hope, easily being able to add various data
visualizations and statistics to any StreamField.  It works by
assembling a list of visualization templates that receive a "data"
context varible, and also displaying the list of registered statistics
functions.  A user can mix and match these two pieces and, when the
page is rendered, the appropriate data will be queried and displayed.

Of note: statistics functions can take arguments, which can be edited
by the user in the "params" field.  Right now, this feature works best
with simple string-like arguments.  When rendering, the value of
"params" is split on whitespace and the results are passed to the
statistics function as its arguments.  If this ends up being
inadequate, it probably wouldn't be that hard to extend this or make
it more user friendly.

Note that you can still include statistics tags in your post content
without using statistics block, but there's not a super
straightforward way to handle the "visualization" aspect of that, so
right now this feature works best with "single number" statistics.

Note: it might make sense to conceptually separate
"statistics" (numbers) from "data series" (lists of data pairs) at
some point.