Separate statistics dataset functions into numbers and maps
Since it's seeming like we will want to use these two types of
functions in different places on the site, it makes sense to
conceptually separate numbers (that return a single number value) from
maps (which return a list of pairs).

This slightly changes the API for defining statistics, but I think
ultimate it makes it better.