Refactor filters
Filters are now programatically created from a list of dicts.
Each dict includes the filter name and type, and some contain the category slug, a modifier for the field lookup, or the choices to validate by.

This allows fields to be added more readily and drastically reduces amount of code needed for filters.