Display DataItems on category page
Note: there is the potential here for an arity mismatch between the
number of params given by the admin and the number expected by the
function.  Ideally, this would be validated, but I couldn't get that
to work so instead I've chosen to just display nothing if we would
encounter an error.