Run tests and retrieve junit results in CI
Added a new configuration that is strictly to be utilized in CI. I dont see any
reason why we should be enabling this in prod. Could have controlled this option
via environment variables but this seemed cleaner.

Updated the CI playbook as well to fetch back the test results from the docker
container.