Skip filtering if daterange would be illegal
If the lower bound exceeds the upper bound, then skip filtering by
dates altogether.  It's possible we want some kind of client-side
helper for the date picker to prevent people from doing this (e.g. an
explicit date range picker, rather than two date pickers) or at least
some kind of validation.

But for now this is the simplest path to preventing code 500 errors
and is in line with how we skip filtering for other kinds of invalid
dates.