Wire up the download data link for the incident filters footer
Remarks:

1. Since the download link depends on the more-or-less absolute path
to an IncidentIndexPage, I couldn't think of a way to do this other
than just.. getting the path to that page in whatever context we're
in (home page, category page, actual incident index page) and passing
that into the IncidentFilter widget JS.

2. There may be a better way to do all of this, in fact, but as far as
I can tell it works.

3. The filters passed to the export view are those that "would be"
applied, not necessarily those that actually are.  That is, if you
change the filters, then click the export link without clicking apply,
the filters applied to the export are the ones you have selected even
though you never "applied" them to the page itself.  This is kind of a
trade off, I guess.  In practice I'm not sure it matters that much?