Chown root owned files
In the previous build iteration for the developer environment, we were running
all the containers through the root user which is a discouraged security
practice. To repair that situation, lets fail the start-up if we detect root
owned files in the directory and provide a shortcut Make command so developers
can rectify that sitaution.