Fix an error where a blank category filter caused an exception
The problem was we were not taking into account what happens if
someone requests a category filter but the filter is empty,
i.e. `?categories=` with nothing else.  And it turns out we have a
handy function already written to convert potentially unsafe user
input into a nice list of integers.