Writes tests for date range filter changes
Adds two tests, one that confirms queries ranges run in December (month
"12") do not raise exceptions, and the second is more of a context test
that honestly should pass all the time (as is the case for all tests
:wink:). If one fails and the other does not, that information will be
useful when debugging in the future.

Not using test parametrization, which would DRY out the tests somewhat,
because we're not currently using pytest for our tests.