Don't ignore errors on createdevdata in CI
We'd moved the `createdevdata` action to the "post" list of django
management tasks in the django-stack role. Errors there are ignored by
default, since those commands often aren't idempotent. It appears now
that the `createdevdata` is indeed idempotent: I was able to run it
multiple times without error on a local container for the "ci" scenario.

Closes #275.