Removes "safety" from dev reqs
The "safety" pip lib should be installed *only* in CI or the developer
environment, *not* in the containers running the application.

The docs were not clean on the update strategy, so I've updated those.
Regenerates all pip requirements after moving the safety requirement to
devops/requirements.txt, via the new Makefile target.