Ensure that incident filters are sorted alphabetically
Previously they were being iterated straight out of a python
dict, but python dicts are not guaranteed an order as of 3.6
and earlier. Since these filters are used as choices on a model
field, this was causing an amusing problem where migrations would
consistently be created for that model every time `makemigrations`
was run, since the choice order was changing every time.

As of 3.7, dictionaries preserve insertion order so, if desired,
we could remove the use of OrderedDict when upgrading to 3.7 and
not mandate sorting alphabetically.

Resolves #649