Prevent pages without templates from raising server errors
OrganizationIndexPage and OrganizationPage never got templates.
We're preserving them as Page subclasses both to avoid a
migration and in case we decide to use them as Pages in the
future, but in the absence of templates for now, this commit
creates fallback behavior that does not require the use of a
template.

OrganizationIndexPage: raise 404
OrganizationPage: redirect to a BlogIndexPage filtered to posts
   from that organization