Ensure categories passed as props are enabled in filtering
If we are viewing a category page, the category is passed to
`IncidentFiltering` as a prop, but this component was only looking for
categories in the URL when determining its state.  This change makes
it so we will look at the props for any categories given whenever
`categoriesEnabled` needs to be determined.

Refs #677