filtering: Truncate date on both sides of the fuzzy date range
Only the month and year of the `date` field matter when finding the
fuzzy date range for an incident whose exact date is unknown, so we
want to truncate the `date` on both ends of the fuzzy range.
Otherwise we end up with a fuzzy range whose end date depends on the
day part of the `date` and our range will be larger than a month.

Fixes #617