Separate replacement logic from appending logic
In my view, appending and replacing new incidents on a page are
sufficiently different in their implementation that fully separating
these functions makes it easier to read and understand what's
happening here.

I've tried to establish some groundwork for building an even more
coherent fetch-transform-splice workflow, which is beyond the scope of
the issue at hand.  But I believe it's possible.