Convert `pk` values to `id` values for incident filtering
The new version of the autocomplete API returns (and expects to be
given) items with an identifier field named `pk` instead of `id`.  The
data in that field is the same format.  So I think the easiest way to
manage this change is, basically, rename the field immediately before
sending it or after receiving it from the API.

There is a lot of code in the Incident Filtering component that relies
on the `id` field, so I think it's easiest to leave that alone and
rename the field only when we need to interact with the API, rather
than renaming the field everywhere, which would be an extensive and
intrusive change.