Do not run queries not supported by `FakeQuerySet` if previewing
If we're looking at a wagtail preview, then Django's queryset is
replaced with a fake queryset object that, presumably, doesn't alter
your data, since it's just a preview.  We like no-op previews.  But
the fake query set supplied by wagtail doesn't know what to do with
some filters and if it encounters them, it will raise an error.

So, wagtail does have a `request.is_preview()` method we can use to
detect if someone is previewing the page or viewing it "for real," and
in the former case not run any unsupported queries.

One thing to note here: `is_preview` is something that wagtail adds to
the request object somewhere in its internal views, as opposed to in
the middleware.  The consequence of this is that I have had to update
some tests that called `get_context` (the method that now relies on
`is_preview`) to tack this attribute onto the request, since those
tests weren't going through the "full stack" of wagtail and thus the
attribute wasn't there.