Make cache invalidation tests more idempotent
So, these tests were failing as a result of some other tests that were
changed, which makes me think there's something going on in the
scaffolding setup for it, or that it's somehow reliant on the DB being
in a certain state prior to running.

Either way, I've copied this configuration over from another set of
tests that creates the site/root page/home page situation needed for
wagtail and its friends to be happy and now these tests pass.