Fix create and preview function for models with template blocks
`StyledTextTemplateBlock` and `RichTextTemplateBlock` both process
rich text via the template engine.  In Wagtail 2, Draftail is the rich
text backend, which can add escaped HTML entities to text intended to
be read as a template tag.  This change unescapes the HTML in tags
belonging to these blocks.

I've also generalized the "unescape within a tag" function so we can
use it where we need it without duplication.

There are new tests for creating and previewing `IncidentPage`s and
`SimplePage`s.