Add new type of incident filter for ManyToMany-Through relations
As far as I can tell, we don't have a way of expressing a filter that
is based on a field related to `IncidentPage` via a `ManyToMany`
relationship with a "through" model.  This is a fairly basic attempt
at providing that feature, but it seems to work in the case of
`IncidentPage.targeted_journalists`.