Add related name for journalist relation on TargetedJournalist
It turns out we actually do need this related name, for when we are
trying to count up how many journalists were targets for a certain set
of incidents.  In that case we have TargetedJournalist objects and we
want to ask for Journalists that belong to that set, so we have to
have this name to follow that "thread."

If this doesn't make any sense, see a future commit that adds the
`num_journalist_targets` statistics tag.