filters: Treat non-integer values for RelationFilter as invalid
Before this change, `RelationFilter` would happily pass along whatever
values it's given directly to the query function.  But we know that
these values have to be integers, since that's what we're querying
for.  I've add a `clean` function for this filter that mimics
`ManyRelationFilter.clean`, only without the need to handle "many"
values.

The effect of this is that cleaning the data for a `RelationFilter`
either returns an `int` or `None` if invalid, or raising an exception
in strict mode.