factories: Enhance common and category factories
Add factories for specific categories we want to create as part of
making a development site.  For this to work we need factories linking
categories to filters so that the search UI will work.

Add a factory for the wagtail `Site` object for use in development.
We need this for the aforementioned search, but also this could be
useful in testing or other places where we want/need to standardize
how this Site is configured.