factories: Enhance home page factory
Adds more representative fake data.  Perhaps more important for
usability is it will (a) have a default title of "Home" and (b) defer
to pre-existing Home Pages with that title.  This means we can just
call `HomePageFactory()` and it will do the right thing in most cases
and we needn't worry about checking if one already exists.