dev data: Enhance createcategories command
There was a lot of stuff in this file that was never updated when we
moved to class based filters.  Now mostly what this file does is defer
to the logic contained in the factories, or create catgories with a
specific set of filters.

As an aside, I actually tried to somehow get the list of filters to be
part of the factories, but you ended up with a lot of one-off classes
like `ArrestFactory`, `EquipmentSearchFactory`, etc., which is
possibly too verbose for me.  But we could go that direction if we
wanted.