factories: Add uniqifying method to certain autogenerated factories
In the case of some of our unit tests, we depend on an exact number of
objects being created.  This is not guaranteed for non-deterministic
factories with `django_get_or_create` applied to unique fields, which
particularly applies to incident-adjacent many-to-one or many-to-many
relationships.

In the case of development data, we would like to retain the
Faker-based non-deterministic fields, but for tests we would like to
have consistency.  I've added a superclass for some of the more
susceptible factories with a trait that can be switched on to override
the Faker-based name or title field with a meaningless but unique
field.  This is fine when, as in our test suite, we don't care about
how realistic something's name is but we do very much care about how
many of them there are.

This change should reduce the number of intermittent test failures
based on Faker data collisions.