Make `HomePageFeature` related to `Page`, generically
Here, we change the foreign key on this model to point to `Page` in
generic, rather than `IncidentPage` in specific.

I'm also taking the opportunity here to clean up this relation a
little bit, such as by making it `not null`.  The template code
actually throws an error if this field is null!  This entity doesn't
make sense without a `page` related to it, so I've made that
impossible.  And correspondingly, removed the allowability of blank
values and the `SET_NULL` behavior.

It's possible this will make the home page execute slightly more
queries, as one other thing we can no longer do is do this
incident-page-specific query:

```
prefetch_related('page__categories__category')
```

Since blog pages don't have categories.  There might not be a way to
easily restore this without fully separating incidents and blog posts
and re-writing a lot of the code downstream from this.  I suspect the
potential performance loss would not be worth doing that for.