Rework home page features query to be preview compatible
I think the last time I changed this code (when we added blog posts to
the types of pages allowed to be featured), I was overambitious with
how efficient I wanted to make this query.  The new query is less
efficient for not using `PageQuerySet.specific()`, since, when
previewing, the set of all `Page` data is not faked, only attributes
on the page actually being previewed.  That is, `self.features` is
faked and properly previews changes, but `Page.objects` is not and
does not.

I did like the old query, but I think we might have to use this new
one if we want previewing to work.  And we can only have 6 features at
most, so even if our total query count goes up it's not by much.