Move recently-updated query definitions to a manager method
This code layout seems nicer to me, as now the filter doesn't "own"
the power of filtering incidents in this way and it can be used more
easily elsewhere.  And we do perform this sort of filtering or have
logic based on how recently an incident was updated in other places on
the site, so I think it might be nice to provide this query for this
use-cases.