Switch the incident page to display workers over targets
In leak cases, we now show the new, more specific classification
Government Worker instead of the old Target information.

It actually looks like there's a change in behavior here (possibly
fixing something unintended) where the link on a target's name now
goes to the search page with the filter for other incidents in which
that target's comms were obtained.  Before, it linked to the filter of
all incidents *targeting* this person *generally*, which is not the
same thing under our schema.  The new behavior seems more correct to
me, as the linked-to filter is now the same context as the link.