tags: Add queryset for unused tags to tag manager
Here we're adding a manager method so you can write

    CommonTag.objects.unused()

And get all the tags without any associated incidents.  I'm choosing
to do this at the manager level since there are potentially multiple
places in the site where we want this data and it would reduce
repetition to have it in a single place like this.