migrations: Partially squash home migrations to avoid error
I've squashed migrations in the `home` app up to and including number
0019 in order to remove a data migration that involved the need to
call `SearchSettings.for_site()`, which now causes an error because:

1. `for_site` is a model method, which you don't have access to unless
you import the model directly, instead of reconstructing it from
`get_model`, which only allows access to fields.

2. If you import a model directly during a migration and you've
changed that model since the migration was made, the migration will
crash.

3. A change has now been made to the `SearchSettings` model.

Since migration 0019 is a data migration marked as `elidable=True`, it
will be removed/elided when squashed.  So I've done this to avoid the
error.  I've only done up through that migration to limit the
invasiveness of the overall operation.