Add missing migration
I'm not sure why this is being generated, but `makemigrations` creates
this file, even though the field did not change in the model.
Possibly a squash bug?  It looks like a default value of `''` somehow
was introduced in the squashed migration file.  This is a little
cluttered, but seems safe to run.