pagination: Gracefully handle noninteger page/endpage values
This fixes a problem where a request with noninteger values for the
`page` and/or `endpage` querystring params would result in a 500
response.  IMO it should not be possible to raise an exception in our
application just by requesting a malformed URL.

I've fixed this by checking both of the page-related request
parameters and, if either one does not parse into an integer, treating
the entire pagination request as incorrect and resetting the page to
1.

I am fairly confident that our front-end pagination code generates the
correct links, so this is really just being done to clean up our error
logs and prevent exceptions for edge-cases.