npm: Require watchpack 1.6.1
It appears something in watchpack 1.7.2 (released 3 days ago) is
tripping up our npm audit checks.  I'm not fully sure what to do about
this, but requiring this specific version seems to cause the chain of
dependencies to be the ones that don't have known vulnerabilities.

I predict we will eventually remove this requirement, or possibly
change our auditing system to skip dev dependencies.