Remove deprecated lifecycle `componentWillReceiveProps`
The Suggestions component was using a deprecated lifecycle hook that
will be fully removed in version 17 of react.

See
https://reactjs.org/blog/2018/03/27/update-on-async-rendering.html#migrating-from-legacy-lifecycles

I've attempted to migrate this component to a new way of doing things
by storing the "mirror" of nextProps.suggestions[index].id in the
state as `currentId`.  I hope nothing broke here, but I played around
a lot with the autocomplete fields in the filters area and everything
seemed to work how I would want it to work, so it's probably okay?