Add `-e` to make command to run tests in CI environment
This ensures that non-zero exit codes during any "intermediate"
command (e.g. in a sequence of semicolon separated commands, which we
have here) are propagated upwards and the entire call to `bash`
returns a non-zero code.  This is needed so that Circle CI becomes
aware that the test command failed.