logging settings: Remove `content` from logged response fields
The content of JSON responses can sometimes be large, we don't want or
need this in our log files.  So this change updates the
`RESPONSE_FIELDS` setting of to be the default list of fields minus
`"content"`.  This will remove the content from what is logged for all
responses, JSON and non-JSON.

Reverts `CONTENT_JSON_ONLY` to its default setting of True.  We can do
this because our goal of excluding content from logs is achieved by
the above.

Also changes the default content encoding to UTF-8.  Even though we
are not including the content, if that ever were to happen, it could
cause an error if the default setting of `'ascii'` is used, since much
of our content is not decodable into ASCII.