Allow topic pages to be featured on the home page
1. Allow topic page models to be chosen as a target of
HomePageFeature.

2. Add markup and styles for a featured topic page grid
item (inheriting much of how the header is displayed on the topic
page, only scaled down to fit on a grid of other feature page boxes).

3. Slightly change how templates are chosen for featured pages.  There
is now no "default" option.  If a featured page is (somehow) not one
of the three allowed choices, nothing will be rendered there.