Rework styles to avoid distorting the image
I probably changed more than was strictly necessary here, but I wanted
to remove the use of absolute positioning, which was causing a problem
when at mobile-widths, I think due to the change of width one of the
outer-level container elements.

I think the result is actually somewhat cleaner.  Though I also
arbitrarily chose a height for the overall block at small sizes, when
it's not being height-constrained by anything else.