admin: Add function to correct misnumbered order <input> fields
Due to what appears to be a bug in wagtail, inline panels with a
`min_num` defined will render the hidden order inputs for `Orderable`
inline models as blank for the first N inputs, where N is the number
defined on the `min_num` field.  Perhaps this is slightly confusing?
See this issue in wagtail for further enlightenment:

https://github.com/wagtail/wagtail/issues/4010

I've added a JS function to the admin pages that inspects the hidden
input fields that wagtail renders to store the order parameter.  If
the value is blank, it corrects that value.  This function was, in
fact, adapted from that very issue linked above.