related incidents: Incorporate location into algorithm
Here I've added a comparison for the (city, state) pair for incidents
when searching for related ones.  This is used as a tie-breaking if
there are too many other incidents with overlapping tags.

I've also simplified the procedure for determining how many related
incidents to collect.  The function now follows these rules:

1. Always return all explicitly defined relationships.  If an editor
has selected incidents as being related to this one, I consider that
an indication that the relationship should always be listed.

2. Limit the number of additional related incidents according to a
`threshold` parameter.  If the number of explicit relationships does
not meet the threshold, search for more (using the above tag/location
algorithm) until we meet that threshold.

3. If we have not met the threshold after exhausting all incidents
with overlapping tags, or matching city and/or state values, then do
not return any more incidents.  Beyond this, I consider the
relationship too tenuous to include.