autocomplete: Fix suggestion-checking for Single autocomplete inputs
What I'm doing here is differentiating between Objects and Arrays when
checking for whether or not we need to re-fetch the suggestion
values.  The previous check was crashing when it processed a value
that was not an array, because it was looking at `value[0]`.

I've tried to be somewhat thorough here, but basically the structure
of this new conditional is: we're trying to check if the component's
`value` prop is missing its `title` attribute, meaning we need to
refresh.  If it's an array, look at the first element and check that
value for a `title`.  If it's an "object" (in this case, an object
literal, something that would be represented in JSON as a string
beginning with "{"), then check that value directly for a `title`
property.  In either case, we want to fetch the values if the `title`
property is absent.