forms: Allow fields to be appended to the email subject
Here's another property on form fields that permits better control of
the email generated on form submission.  This time, it's a check box
that specifies if the field should be included in the subject of the
email.  As many fields as are desired can be given this property and
they will all be included in the subject.