incident: Add JSON endpoint to query summary data
We are using this data for a feature in the admin, and we require
access to it from Javascript.  So this change adds a `/summary` API
endpoint on `IncidentIndexPage` models which parses the same search
query parameters as the base page view, except it only returns the
summary data for journalists affected, institutions affected, and
total incident count.