Add AlignedCaptionedImageBlock to choices for Incident body fields
The "base" image block does not allow captioning or the automatic
display of the image attribution, which is something we want for
images displayed in the bodies of incidents.  Essentially we want
these to work the same way as they do in other parts of the site.

I did not remove the old image field, since we probably have older
images in the production DB using that block and I'm not sure we're
ready to potentially cause problems there.