Remove stats-file caching from webpack loader
I'm not sure exactly at what point this bug started happening, but
locally and on CI, I've been encountering a bug where the webpack
loader would try to load a bundle while webpack was compiling and
cache the stats file then, at which point the file contains only

```
{"status": "compiling"}
```

and proceed to either timeout (or hang if not timeout configured)
because the file was cached in that state.

Disabling caching in dev mode fixes the problem.