Do not override TooltipEntity's children
I think this fixes a problem where the entity in the admin text-box
would have odd behavior, e.g. making text insertion and deletion occur
not where expected by the cursor location.  It's possible that having
this be there made the entity take up a different amount of "text
space" in the editor state than it really did.  I'm not entirely sure.