Only run DB query for settings if needed
Having the `get_settings` statement at the top-level of this template
means we query the database to get the settings on every incident
being rendered, which is not necessary.

The underlying problem here is we are using this same template for
displaying a single incident's information and also displaying a
"teaser" for many incidents.  For a single incident, the settings call
is probably fine (and not a huge performance drain), but for
many (i.e. if `teaser == True`) then we want to avoid that.