Remove extraneous N+1 query for incident updates
`incident.recently_updated` makes a DB query for the most-recently
made incident update model.  We actually already have a queryset
method that efficiently gets this information for an entire set of
Incident pages, so let's use that.  This way we avoid running one
query per-incident being displayed in a list.

Note: this is only really a partial solution.  If there *is* an update
in the past 7 days, the template will call `last_updated`, which will
run another query.  Ideally, we could prefetch this information, too.