Improve targeted journalist/institution summary query
This change replaces our use of `total_queryset` with `incident_ids`,
i.e. replacing an entire queryset (which has the potential to contain
some complex filters) with a list of IDs, which we query once and then
save as a Python value.  This causes the Django ORM to use an
`IN (<static list of ids>)` query rather than `IN (potentially
complicated subquery)`.

In particular I think this saves us a lot of trouble on the
institution summary, which Django was running as a left join with two
very large `OR` expressions in the where-clause.  But all the
complexity in those expressions was due to the filtering, which we can
change to a static list of IDs and get the same results with a lot
less number crunching.