topic page: Better-optimized incident categorization query
This change removes a somewhat large-ish subquery from the WHERE
clause of this Category Page query.  The reason we were using a
subquery in the first place (I think) was that we wanted the database
to be able to apply the `LIMIT` to the number of incidents per
category.  Django, however, is not able place a limit (expressed in
python as a slice) to a queryset a part of a prefetch without a
subquery.  The actual SQL generated by this is somewhat ugly, so this
change puts the limit calculation into our python code, via the
marshmallow schema's context dictionary.  This allows the actual query
to be much simpler and more performant.