incident page: Add `unique_date` field for cursor-based pagination
The most performant method of pagination that Django Rest Framework
offers is cursor-based, but that requires a special field to order by
that meets certain parameters.  See

https://www.django-rest-framework.org/api-guide/pagination/#cursorpagination

for more details.  The main points are that the field is a unique,
non-null string that also reflects the overall order want to view the
results in.  IMO the simplest way to go here is to add the page's
unique key onto the date field.  The date is how we want to order the
results, adding the key gives it the properties that make cursor
pagination useful.

I've set up the migration here to auto-populate the field for exising
incident pages, and also adjusted the model to compute the field
whenever a new one is saved (this also catches the case of when the
date field is changed).