Allow rich-text statistics widget to be used for stat boxes
Stat boxes formerly stored their "value" as a char field, with a text
entry widget.  We commonly use them for statistics tags, so it would
be nice to be able to use our more user-friendly "numincidents" widget
with them.  This change updates stat boxes to store their "value" as
rich text (with a very limited feature set, including the incidents
widget) and display the value inline (without block-level tags).