Add fuzzy date logic to topic page date filtering
Here we apply the same logic that we use on the search page, where an
incident with exact date unknown is said to span the entire month that
contains its specified date.  The implementation is slightly trickier
as our use of the ORM here is tricker than in the search code.  I'm
pretty happy with this `Transform` technique, though.