Use an explicit seed for reproducable randomness in the test suite
Adds a test runner whose purpose is to configure the seed for Factory
Boy (as described here
https://factoryboy.readthedocs.io/en/stable/recipes.html#recipe-random-management)
to prevent randomly generated objects from causing unexpected and
ephemeral test failures.

This is particularly common for our incident-related inline objects
that consist solely of a single field that must be unique.  When
randomly generating several of these for a test, if the same value is
chosen for that field for more than one object (an uncommon but
definitely possible occurrence) then only a single object will be
created where the test expected two, thus potentially causing a test
assertion against an aggregate count to fail.

My hope is that using the same seed locally and on CI will prevent us
from being surprised when such problems occur.