Change "wait on element" logic to "wait on URL" logic
Experienced some transitory errors on this test in the CI environment
where the `wait` call would fail with:

```
selenium.common.exceptions.StaleElementReferenceException: Message: The element reference of <a class="button button--outline button--center js-incident-loading-next-link" href="?page=2"> is stale; either the element is no longer attached to the DOM, it is not in the current frame context, or the document has been refreshed
```

I've attempted to sidestep this problem by replacing the "wait for an
element to exhibit a certain state" call with "wait for the URL to be
changed" call.  The URL will always change if the fetch-incidents
request succeeds, so I hope we can reliably use that as the "go ahead"
to click the button again without worrying about it going
stale (since, presumably, the URL cannot go stale).