Fix a bug in generating unique dates
Previously we were generating unique dates for incidents by concatenating
incident date with incident database id on save. Unfortunately, because
when an incident is first saved it has no database id, this was
generating unique date keys that looked like "2020-06-16-None" and were
not guaranteed to be unique.

The test was missing this error because using the IncidentPageFactory
triggered multiple calls to .save() and the key was corrected on the
second call--thus the test passed.

I've rewritten the functionality to use a generated-on-the-fly uuid
instead of database id and the test to compare two incidents against
each other to verify their unique dates are not identical. I also
rewrote the test to use IncidentPage directly instead of the factory.